# evidenciesib-api-externa/openapi/openapi_comanda.json
{
  "openapi" : "3.0.1",
  "servers" : [ {
    "url" : "/evidenciesibapi/externa"
  }, {
    "url" : "http://localhost:8080/evidenciesibapi/externa"
  }, {
    "url" : "https://dev.caib.es/evidenciesibapi/externa"
  }, {
    "url" : "https://proves.caib.es/evidenciesibapi/externa"
  }, {
    "url" : "https://se.caib.es/evidenciesibapi/externa"
  }, {
    "url" : "https://www.caib.es/evidenciesibapi/externa"
  } ],
  "paths" : {
    "/secure/estadistiques/v1" : {
      "get" : {
        "operationId" : "estadistiques",
        "responses" : {
          "default" : {
            "description" : "default response",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/RegistresEstadistics"
                }
              }
            }
          }
        },
        "security" : [ {
          "BasicAuth" : [ ]
        } ]
      }
    },
...
  },
  "components" : {
    "schemas" : {
      "Dimensio" : {
        "type" : "object",
        "properties" : {
          "codi" : {
            "type" : "string"
          },
          "valor" : {
            "type" : "string"
          }
        }
      },
      "Fet" : {
        "type" : "object",
        "properties" : {
          "codi" : {
            "type" : "string"
          },
          "valor" : {
            "type" : "number",
            "format" : "double"
          }
        }
      },
      "RegistreEstadistic" : {
        "type" : "object",
        "properties" : {
          "dimensions" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/Dimensio"
            }
          },
          "fets" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/Fet"
            }
          }
        }
      },
      "RegistresEstadistics" : {
        "required" : [ "temps" ],
        "type" : "object",
        "properties" : {
          "temps" : {
            "type" : "string",
            "format" : "date-time"
          },
          "fets" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/RegistreEstadistic"
            }
          }
        }
      },
...
    },
    "securitySchemes" : {
      "BasicAuth" : {
        "type" : "http",
        "scheme" : "basic"
      }
    }
  }
}